Objective working
It's now possible to walk to the other side of the map, pick up an object and walk back and when colliding with the drop off point the game is stopped. if the player walks into the drop of point without fist colliding with the pick up point the game will do nothing. also the pickup point becomes invisible after it has been touched